Generated ...ImlpBase class uses suspend funs
Rather than having the user code which implements the server return Defered<T>, if can now implement a suspending function which simply returns T. This results in more straight forward user code on the server side; the {{methodName}}Internal function calls the method in an async { } block to keep most of the other generated code the same. This is likely not idea.